+++ lib/calendav/clients/events_client.rb
@@ -20,12 +20,13 @@ module Calendav
         endpoint.delete(url: event_url)
       end
 
-      def list(calendar_url, from:, to:)
+      def list(calendar_url, from: nil, to: nil)
         request = Requests::ListEvents.call(from: from, to: to)
 
         endpoint
           .report(request.to_xml, url: calendar_url, depth: 1)
           .xpath(".//dav:response")
+          .reject { |node| node.xpath(".//caldav:calendar-data").text.empty? }
           .collect { |node| Event.from_xml(calendar_url, node) }
       end