+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Support creating and deleting calendars.
 * Updating of events.
+* List all events on a calendar (no timespan required).
 
 ## [0.0.1] - 2021-06-13